download-js-corpus: Now remove dirs that are cheating (aka somewhat non human readable code idk) with u.
